+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 1. [](#bugfix)
     * Fixed custom actions not working
     * Fixed custom folder in `mediapicker` field not working
+    * Fixed export title when not using CVS [#51](https://github.com/trilbymedia/grav-plugin-flex-objects/issues/51)
 
 # v1.0.0-rc.9
 ## 03/20/2020