+++ CHANGELOG.md
@@ -5,7 +5,7 @@
     * Fixed custom actions not working
     * Fixed custom folder in `mediapicker` field not working
     * Fixed export title when not using CVS [#51](https://github.com/trilbymedia/grav-plugin-flex-objects/issues/51)
-    * Fixed preview in Page list view [#1845](https://github.com/getgrav/grav-plugin-admin/issues/1845)
+    * Fixed preview in Page list view [admin#1845](https://github.com/getgrav/grav-plugin-admin/issues/1845)
 
 # v1.0.0-rc.9
 ## 03/20/2020